comimt2

+++ ProyectoEjemplo/src/main/java/com/proyecto/servicio/EstacionServiceImpl.java
@@ -2,8 +2,10 @@ package com.proyecto.servicio;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -14,7 +16,7 @@ import com.proyecto.persistencia.EstacionRepo;
 
 @Service
 public class EstacionServiceImpl implements EstacionService {
-	
+
 	@Autowired
 	private EstacionRepo estacionRepo;
 
@@ -24,19 +26,25 @@ public class EstacionServiceImpl implements EstacionService {
 	}
 
 	@Override
-	public Estacion verEstacion(int id) {
-		return estacionRepo.findById(id).get();
+	public Optional<Estacion> verEstacion(int id) {
+		return estacionRepo.findById(id);
 	}
 
 	@Override
-	public List<Estacion> eliminarEstacion(int id) {
-		estacionRepo.deleteById(id);
-		return estacionRepo.findAll();
+	public void eliminarEstacion(int id) throws NoSuchElementException {
+		if (estacionRepo.findById(id).get() != null) {
+			estacionRepo.deleteById(id);
+		}
 	}
 
 	@Override
 	public List<Estacion> listaEstaciones() {
 		return estacionRepo.findAll();
+	}
+
+	@Override
+	public Estacion actualizarEstacion(Estacion estacion) {
+		return estacionRepo.save(estacion);
 	}
 
 }
